backend/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.21
scikit-learn>=1.2
numpy>=1.23
pillow>=9.0
pytesseract>=0.3
easyocr>=1.7
opencv-python-headless>=4.7
matplotlib>=3.7
python-multipart>=0.0.6
joblib>=1.2
pytest>=7.0
requests>=2.28
httpx>=0.24
